# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: SSCMS API
  description: 'Smart Student Club Management System -- Backend Sprint 1 (Milestone
    3) covered Epics 1-3: Role-Based Portal Access & Onboarding, Venue Booking & Asset
    Management, and Event Lifecycle & Approval Workflows. Backend Sprint 2 (Milestone
    4) adds Epics 4-6: Task Management & Volunteer Workflows, QR Attendance & Live
    Execution, and Certification & Budget Analytics.'
  version: 1.1.0
paths:
  /auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '409':
          description: student_id or email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/me:
    get:
      tags:
      - Authentication
      summary: Read Current User
      operationId: read_current_user_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
      security:
      - OAuth2PasswordBearer: []
  /auth/change-password:
    post:
      tags:
      - Authentication
      summary: Change Password
      operationId: change_password_auth_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /auth/bulk-import:
    post:
      tags:
      - Authentication
      summary: Bulk Import Users Auth Alias
      operationId: bulk_import_users_auth_alias_auth_bulk_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_bulk_import_users_auth_alias_auth_bulk_import_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkUserImportOut'
        '403':
          description: Faculty Coordinator role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /clubs/{club_id}/chat:
    get:
      tags:
      - Club Chat
      summary: List Chat Messages
      operationId: list_chat_messages_clubs__club_id__chat_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: club_id
        in: path
        required: true
        schema:
          type: integer
          title: Club Id
      - name: student_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Student Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessageOut'
                title: Response List Chat Messages Clubs  Club Id  Chat Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Club Chat
      summary: Send Chat Message
      operationId: send_chat_message_clubs__club_id__chat_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: club_id
        in: path
        required: true
        schema:
          type: integer
          title: Club Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageOut'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /events/{event_id}/finalize:
    post:
      tags:
      - Certification & Budget Analytics
      summary: Finalize Event
      operationId: finalize_event_events__event_id__finalize_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeEventOut'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: Event is already finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /certificates:
    get:
      tags:
      - Certification & Budget Analytics
      summary: List My Certificates
      operationId: list_my_certificates_certificates_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CertificateOut'
                type: array
                title: Response List My Certificates Certificates Get
      security:
      - OAuth2PasswordBearer: []
  /analytics/budget-overview:
    get:
      tags:
      - Certification & Budget Analytics
      summary: Budget Overview
      operationId: budget_overview_analytics_budget_overview_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetOverviewOut'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    Body_bulk_import_users_auth_alias_auth_bulk_import_post:
      properties:
        file:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: File
        payload:
          anyOf:
          - items:
              $ref: '#/components/schemas/BulkUserImportRow'
            type: array
          - type: 'null'
          title: Payload
      type: object
      title: Body_bulk_import_users_auth_alias_auth_bulk_import_post
...
    BudgetOverviewOut:
      properties:
        clubs:
          items:
            $ref: '#/components/schemas/ClubBudgetOverview'
          type: array
          title: Clubs
        total_allocated:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Allocated
        total_spent:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Spent
        active_bookings:
          type: integer
          title: Active Bookings
        inventory_on_loan:
          type: integer
          title: Inventory On Loan
      type: object
      required:
      - clubs
      - total_allocated
      - total_spent
      - active_bookings
      - inventory_on_loan
      title: BudgetOverviewOut
...
    BulkUserImportOut:
      properties:
        total:
          type: integer
          title: Total
        imported:
          type: integer
          title: Imported
        failed:
          type: integer
          title: Failed
        results:
          items:
            $ref: '#/components/schemas/BulkUserImportRowResult'
          type: array
          title: Results
      type: object
      required:
      - total
      - imported
      - failed
      - results
      title: BulkUserImportOut
    BulkUserImportRow:
      properties:
        student_id:
          type: string
          title: Student Id
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
          default: student@123
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        year:
          anyOf:
          - type: string
          - type: 'null'
          title: Year
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
          default: Student
        club_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Club Id
      type: object
      required:
      - student_id
      - name
      - email
      title: BulkUserImportRow
    BulkUserImportRowResult:
      properties:
        row:
          type: integer
          title: Row
        email:
          type: string
          title: Email
        status:
          type: string
          title: Status
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - row
      - email
      - status
      title: BulkUserImportRowResult
...
    CertificateOut:
      properties:
        id:
          type: integer
          title: Id
        event_id:
          type: integer
          title: Event Id
        user_id:
          type: integer
          title: User Id
        reason:
          type: string
          title: Reason
        issued_at:
          type: string
          format: date-time
          title: Issued At
      type: object
      required:
      - id
      - event_id
      - user_id
      - reason
      - issued_at
      title: CertificateOut
    ChatMessageCreate:
      properties:
        student_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Student Id
        text:
          type: string
          title: Text
      type: object
      required:
      - text
      title: ChatMessageCreate
    ChatMessageOut:
      properties:
        id:
          type: integer
          title: Id
        club_id:
          type: integer
          title: Club Id
        student_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Student Id
        sender_id:
          type: integer
          title: Sender Id
        sender_role:
          type: string
          title: Sender Role
        text:
          type: string
          title: Text
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - club_id
      - sender_id
      - sender_role
      - text
      - created_at
      title: ChatMessageOut
    ClubBudgetOverview:
      properties:
        club_id:
          type: integer
          title: Club Id
        club_name:
          type: string
          title: Club Name
        allocated:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Allocated
        spent:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Spent
      type: object
      required:
      - club_id
      - club_name
      - allocated
      - spent
      title: ClubBudgetOverview
...
    ErrorOut:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorOut
...
    EventOut:
      properties:
        id:
          type: integer
          title: Id
        proposal_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Proposal Id
        club_id:
          type: integer
          title: Club Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        venue_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Venue Id
        event_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Event Date
        registration_deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Registration Deadline
        published:
          type: boolean
          title: Published
        finalized:
          type: boolean
          title: Finalized
      type: object
      required:
      - id
      - club_id
      - name
      - published
      - finalized
      title: EventOut
...
    FinalizeEventOut:
      properties:
        event:
          $ref: '#/components/schemas/EventOut'
        certificates_issued:
          items:
            $ref: '#/components/schemas/CertificateOut'
          type: array
          title: Certificates Issued
      type: object
      required:
      - event
      - certificates_issued
      title: FinalizeEventOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RoleOut:
      properties:
        role:
          type: string
          title: Role
        club_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Club Id
      type: object
      required:
      - role
      title: RoleOut
...
    TokenOut:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - user
      title: TokenOut
    ChangePassword:
      properties:
        current_password:
          type: string
          title: Current Password
        new_password:
          type: string
          minLength: 6
          title: New Password
      type: object
      required:
      - current_password
      - new_password
      title: ChangePassword
    UserLogin:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: UserLogin
    UserOut:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: string
          title: Student Id
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        year:
          anyOf:
          - type: string
          - type: 'null'
          title: Year
        roles:
          items:
            $ref: '#/components/schemas/RoleOut'
          type: array
          title: Roles
          default: []
      type: object
      required:
      - id
      - student_id
      - name
      - email
      title: UserOut
    UserRegister:
      properties:
        student_id:
          type: string
          title: Student Id
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 6
          title: Password
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        year:
          anyOf:
          - type: string
          - type: 'null'
          title: Year
        role:
          type: string
          title: Role
          description: 'One of: Student, Volunteer, ClubHead, ClubPresident, FacultyCoordinator'
        club_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Club Id
          description: Required for ClubHead/ClubPresident roles
      type: object
      required:
      - student_id
      - name
      - email
      - password
      - role
      title: UserRegister
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login